+++ CHANGELOG.md
@@ -5,6 +5,8 @@ unreleased
 ---------
 - Update braintree-web to v3.51.0
   - Update songbird.js script urls for 3D Secure
+- Default 3D Secure ACS Window Size to `03` (see [`acsWindowSize` option](https://braintree.github.io/braintree-web/current/ThreeDSecure.html#verifyCard))
+- Scope full screen 3D Secure modal to screen sizes with heights of 700px and smaller
 
 1.20.0
 ------