+++ CHANGELOG.md
@@ -4,8 +4,9 @@ CHANGELOG
 unreleased
 ---------
 - Update event-emitter to v0.3.0
-- Update braintree-web to v3.51.0
+- Update braintree-web to v3.52.0
   - Update songbird.js script urls for 3D Secure
+  - Fix issue where bin was not being passed to underlying cardinal SDK
 
 1.20.0
 ------